create reducer and actions using createSlice method
