+++ src/index.js
@@ -1,16 +1,16 @@
-import store from "./store";
-import { addTask, completedTask, removeTask } from "./store/configureStore";
+import store from "./store/configureStore";
+import { addTask, completedTask, removeTask } from "./store/tasks";
 
-const unsubscribe = store.subscribe(() => {
-  console.log("Updated State", store.getState());
+store.subscribe(() => {
+  console.log(store.getState());
 });
 
 store.dispatch(addTask({ task: "Task 1" }));
+
 store.dispatch(addTask({ task: "Task 2" }));
 
-store.dispatch(completedTask({ id: 2 }));
+store.dispatch(completedTask({ id: 1 }));
 
-unsubscribe();
+store.dispatch(addTask({ task: "Task 3" }));
 
-store.dispatch(removeTask({ id: 1 }));
-console.log("Normal Console Log: ", store.getState());
+store.dispatch(removeTask({ id: 2 }));
