--- src/store.js
@@ -1,5 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import reducer from "./store/configureStore";
-
-const store = configureStore({ reducer });
-export default store;
