+++ src/store/configureStore.js
@@ -1,28 +1,6 @@
-import { createAction, createReducer } from "@reduxjs/toolkit";
+import { configureStore } from "@reduxjs/toolkit";
+import reducer from "./tasks";
 
-// Actions
-export const addTask = createAction("ADD_TASK");
-export const removeTask = createAction("REMOVE_TASK");
-export const completedTask = createAction("COMPLETE_TASK");
+const store = configureStore({ reducer });
 
-// Reducer
-let id = 0;
-
-export default createReducer([], (builder) => {
-  builder
-    .addCase(addTask.type, (state, action) => {
-      state.push({
-        id: ++id,
-        task: action.payload.task,
-        completed: false,
-      });
-    })
-    .addCase(removeTask.type, (state, action) => {
-      const index = state.findIndex((task) => task.id === action.payload.id);
-      state.splice(index, 1);
-    })
-    .addCase(completedTask.type, (state, action) => {
-      const index = state.findIndex((task) => task.id === action.payload.id);
-      state[index].completed = true;
-    });
-});
+export default store;
